refactor(runtime): share analyzeImportedModDifference and proxyGuardOnlyEsm
